Feature develop

See merge request wangxiao_xueyanandroid_business/livevideo!26

+++ src/main/java/com/xueersi/parentsmeeting/modules/livevideo/business/courseware/CoursewarePreload.java
@@ -105,6 +105,9 @@ public class CoursewarePreload {
             @Override
             public void run() {
                 logger.i("start delete file");
+                if (file == null || file.listFiles() == null) {
+                    return;
+                }
                 for (File itemFile : file.listFiles()) {
                     if (isCoursewareDir(itemFile.getName()) && !itemFile.getName().equals(today)) {
                         if (!itemFile.isDirectory()) {
@@ -129,6 +132,9 @@ public class CoursewarePreload {
     }
 
     private void deleteFor(final File file) {
+        if (file == null || file.listFiles() == null) {
+            return;
+        }
         for (File itemFile : file.listFiles()) {
             if (!itemFile.isDirectory()) {
                 itemFile.delete();
